+++ packages/near-api-ts/universal/src/client/methods/transaction/getTransactionResult/handleRpcResult/_common/getConversionError.ts
@@ -1,0 +1,5 @@
+import type { InvalidTxError } from '@near-js/jsonrpc-types';
+
+// Right now we aren't sure if it's really possible to meet a conversion error on the chain,
+// and which union options are real, so we return the error as is
+export const getConversionError = (invalidTxError: InvalidTxError): unknown => invalidTxError;