+++ packages/near-api-ts/universal/src/client/methods/transaction/getTransactionResult/handleRpcResult/_common/getProcessingSteps/getNonConversionSteps/refundSteps/getRefundSteps.ts
@@ -21,7 +21,7 @@ const getRefundStepResult = (status: RpcReceiptOutcome['outcome']['status']): Re
   ) {
     return {
       status: 'Error',
-      error: { kind: 'Receiver.NotFound' },
+      error: { kind: 'Receiver.NotFound', context: null },
     };
   }