+++ packages/near-api-ts/universal/tests/integrations/signAndSendTransaction/errors/execution/functionCall/functionCall.test.ts
@@ -4,10 +4,11 @@ import { type Client, keyPair } from '../../../../../../index';
 import type { KeyPair } from '../../../../../../types/_common/keyPairs/keyPair';
 import { createDefaultClient } from '../../../../../utils/common';
 import { startSandbox } from '../../../../../utils/sandbox/startSandbox';
-import { compilationFailed } from './compilationFailed';
-import { executionErrorNew } from './executionErrorNew';
-import { executionErrorOld } from './executionErrorOld';
-import { functionInvalidSignature } from './functionInvalidSignature';
+import { compilationFailed } from './compilationFailed/compilationFailed';
+import { methodInvalidSignature } from './compilationFailed/methodInvalidSignature';
+import { executionError } from './executionFailed/executionError';
+import { hostErrorGuestPanic } from './executionFailed/hostErrorGuestPanic';
+import { deleteActionMustBeFinal } from './executionFailed/newReceiptValidationError/deleteActionMustBeFinal';
 import { functionNotFound } from './functionNotFound';
 import { wasmNotFound } from './wasmNotFound';
 
@@ -28,9 +29,20 @@ describe('signAndSendTransaction › FunctionCall.* errors', () => {
   });
 
   it('return Action.FunctionCall.Wasm.NotFound', wasmNotFound(context));
-  it('return Action.FunctionCall.Compilation.Failed', compilationFailed(context));
   it('return Action.FunctionCall.Function.NotFound', functionNotFound(context));
-  it('return Action.FunctionCall.Function.InvalidSignature', functionInvalidSignature(context));
-  it('return Action.FunctionCall.Execution.Failed for old HostError', executionErrorOld());
-  it('return Action.FunctionCall.Execution.Failed', executionErrorNew(context));
+  it('return Action.FunctionCall.Compilation.Failed', compilationFailed(context));
+  it(
+    'return Action.FunctionCall.Compilation.Failed: MethodInvalidSignature',
+    methodInvalidSignature(context),
+  );
+  it('return Action.FunctionCall.Execution.Failed', executionError(context));
+  it(
+    'return Action.FunctionCall.Execution.Failed: old HostError.GuestPanic',
+    hostErrorGuestPanic(),
+  );
+  it(
+    'return Action.FunctionCall.Execution.Failed: ' +
+      'old NewReceiptValidationError.ActionsValidation.DeleteActionMustBeFinal',
+    deleteActionMustBeFinal(context),
+  );
 });