+++ packages/near-api-ts/universal/types/_common/transactionDetails/processingSteps/refundStep.ts
@@ -9,6 +9,7 @@ export type RefundStepResult =
       status: 'Error';
       error: {
         kind: 'Receiver.NotFound';
+        context: null;
       };
     };